Move gear execution to common/src/python
